# api/openapi.yaml
openapi: 3.0.3
info:
  title: Plex2Jellyfin API
  description: API for managing media library cleanup operations
  version: 2.0.0
  license:
    name: GPL-3.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /webhooks/jellyfin:
    post:
      operationId: jellyfinWebhook
      summary: Receive Jellyfin webhook events
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Event accepted
        '400':
          description: Invalid payload
        '401':
          description: Invalid or missing webhook secret

  # ============ SETTINGS ============
...
  /settings/paths/{kind}:
    parameters:
      - name: kind
        in: path
        required: true
        schema:
          type: string
          enum: [movies, tv]
    get:
      operationId: getSettingsPaths
      summary: List watch paths
      tags: [Settings]
      responses:
        '200':
          description: Watch paths
    post:
      operationId: addSettingsPath
      summary: Add a watch path
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Added
    put:
      operationId: replaceSettingsPaths
      summary: Replace watch paths
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Replaced
  /settings/paths/{kind}/{index}:
    delete:
      operationId: removeSettingsPath
      summary: Remove a watch path
      tags: [Settings]
      parameters:
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [movies, tv]
        - name: index
          in: path
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Removed
...